PHREEQC should now have no USE_MPI
Also no MERGE_FILES in PHREEQC

Not sure about deallocates in worker_closef

git-svn-id: svn://136.177.114.72/svn_GW/phreeqc3/trunk@7045 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ phreeqc/input.cpp
@@ -107,7 +107,7 @@ get_line(void)
 	size_t l = (l1 > l2) ? l1 : l2;
 	if (l >= (size_t) max_line)
 	{
-		max_line = l * 2;
+		max_line = (int) l * 2;
 		line_save =	(char *) PHRQ_realloc(line_save,
 			(size_t) max_line * sizeof(char));
 		if (line_save == NULL)
